requirements.txt
aiofiles==0.6.0
aiohttp==3.7.3
async-lru==1.0.2
async-timeout==3.0.1
attrs==20.3.0
certifi==2020.12.5
chardet==3.0.4
cssselect2==0.4.1
gitdb==4.0.5
GitPython==3.1.12
humanize==3.2.0
idna==2.10
lxml==4.6.2
multidict==5.1.0
Pillow==8.1.0
prettytable==2.0.0
psutil==5.8.0
pyaes==1.6.1
Pyrogram==1.1.13
PySocks==1.7.1
python-magic==0.4.18
reportlab==3.5.59
requests==2.25.1
smmap==3.0.5
svglib==1.0.1
TgCrypto==1.2.2
thisapidoesnotexist==0.2
tinycss2==1.1.0
typing-extensions==3.7.4.3
urllib3==1.26.3
wcwidth==0.2.5
webencodings==0.5.1
yarl==1.6.3
